# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "E-commerce backend API for household chemicals, cosmetics, and home goods.",
        "title": "Commerce Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all orders with filtering and pagination (admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (new, confirmed, shipped, delivered, cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by type (delivery, pickup)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by phone",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name or order number",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.Order"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.Pagination"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get order with items by ID (admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.OrderWithItems"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/orders/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Change order status (admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.UpdateOrderStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "description": "Get order history by phone number (public)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders by phone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone number",
                        "name": "phone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.Order"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new order (public)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.OrderWithItems"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_meransoft_commerce-backend_internal_models.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_meransoft_commerce-backend_internal_models.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiError"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.Pagination"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.CartItemRequest": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.CreateOrderRequest": {
            "type": "object",
            "required": [
                "delivery_date",
                "full_name",
                "items",
                "phone",
                "time_slot",
                "type"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "delivery_date": {
                    "type": "string"
                },
                "delivery_zone_id": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.CartItemRequest"
                    }
                },
                "note": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "time_slot": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "delivery",
                        "pickup"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.OrderType"
                        }
                    ]
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.Order": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivery_date": {
                    "type": "string"
                },
                "delivery_fee": {
                    "type": "number"
                },
                "delivery_zone_id": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "order_number": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.OrderStatus"
                },
                "subtotal": {
                    "type": "number"
                },
                "time_slot": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "type": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.OrderType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_meransoft_commerce-backend_internal_models.OrderItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_image_url": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                }
            }
        },
        "github_com_meransoft_commerce-backend_internal_models.OrderStatus": {
            "type": "string",
            "enum": [
                "new",
                "confirmed",
                "shipped",
                "delivered",
                "cancelled"
            ],
            "x-enum-varnames": [
                "OrderStatusNew",
                "OrderStatusConfirmed",
                "OrderStatusShipped",
                "OrderStatusDelivered",
                "OrderStatusCancelled"
            ]
        },
        "github_com_meransoft_commerce-backend_internal_models.OrderType": {
            "type": "string",
            "enum": [
                "delivery",
                "pickup"
            ],
            "x-enum-varnames": [
                "OrderTypeDelivery",
                "OrderTypePickup"
            ]
        },
        "github_com_meransoft_commerce-backend_internal_models.OrderWithItems": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivery_date": {
                    "type": "string"
                },
                "delivery_fee": {
                    "type": "number"
                },
                "delivery_zone_id": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.OrderItem"
                    }
                },
                "note": {
                    "type": "string"
                },
                "order_number": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.OrderStatus"
                },
                "subtotal": {
                    "type": "number"
                },
                "time_slot": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "type": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.OrderType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_meransoft_commerce-backend_internal_models.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.UpdateOrderStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "enum": [
                        "new",
                        "confirmed",
                        "shipped",
                        "delivered",
                        "cancelled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.OrderStatus"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}